Refactor icon folder
