+++ src/components/common/Icon/IconButton/IconButton.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import Icon from "../Icon";
+
+const IconButton = ({ icon, iconColor, action}) => (
+  <button
+    className='odu-icon-button align-self-center'
+    onClick={action}
+  >
+    <Icon icon={icon} iconColor={iconColor}/>
+  </button>
+);
+
+export default IconButton;
