+++ src/components/common/Icon/index.js
@@ -1,0 +1,9 @@
+import Icon from './Icon'
+import IconButton from './IconButton/IconButton'
+import IconWithText from './IconWithText/IconWithText'
+
+export {
+    Icon,
+    IconButton,
+    IconWithText
+}
