+++ src/components/common/index.js
@@ -1,9 +1,13 @@
-import Icon from './Icon/Icon';
-import IconButton from './IconButton/IconButton';
-import IconWithText from './IconWithText/IconWithText';
+import { Icon, IconButton, IconWithText } from './Icon';
 import Search from './Search/Search';
 import TermCard from './TermCard/TermCard';
 import Markdown from './Markdown/Markdown';
 
-
-export { Icon, IconButton, IconWithText, Search, TermCard, Markdown };
+export {
+  Icon,
+  IconButton,
+  IconWithText,
+  Search,
+  TermCard,
+  Markdown
+};
